Sycning from original

+++ scripts/stop.sh
@@ -2,7 +2,17 @@
 
 set -e
 
-pkill -f constellation-node
+CONSTELLATION=${ENABLE_CONSTELLATION:-}
+
+mapfile -t NODE_TYPE <~/alastria/data/NODE_TYPE
+
+if [ "$NODE_TYPE" == "general" ] && [ ! -z "$CONSTELLATION" ]; then
+    pkill -f constellation-node
+fi
+
 pkill -f geth
+
+# If monitor is enabled, never stops.
+./monitor.sh stop
 
 set +e
